requirements.txt
discord.py==2.3.2
file_read_backwards==2.0.0
Pillow==9.2.0
python-dotenv==0.21.0
rcon==2.1.1
tabulate==0.8.10
python-steam-api==2.0
